feat(alpha-app-provider): change to use ThemeProvider internally
