+++ packages/bezier-react/src/providers/AlphaAppProvider.tsx
@@ -1,18 +1,18 @@
-import React, { useEffect } from 'react'
+import React from 'react'
 
 import {
   type Feature,
   FeatureProvider,
 } from '~/src/features'
-import {
-  type ThemeName,
-  TokenProvider,
-} from '~/src/hooks/useToken'
 import { window as defaultWindow } from '~/src/utils/dom'
 
 import { TooltipProvider } from '~/src/components/Tooltip'
 
-import WindowProvider from './WindowProvider'
+import {
+  type ThemeName,
+  ThemeProvider,
+} from './ThemeProvider'
+import { WindowProvider } from './WindowProvider'
 
 export interface AlphaAppProviderProps {
   children: React.ReactNode
@@ -31,11 +31,6 @@ export interface AlphaAppProviderProps {
    * @default window
    */
   window?: Window
-  /**
-   * Root element to apply theme data attribute to. window of default value is a window object passed by `window` prop
-   * @default window.document.body
-   */
-  rootElement?: HTMLElement
 }
 
 /**
@@ -63,28 +58,18 @@ export function AlphaAppProvider({
   themeName = 'light',
   features = [],
   window = defaultWindow,
-  rootElement = window.document.body,
 }: AlphaAppProviderProps) {
-  useEffect(function updateThemeDataAttribute() {
-    // TODO: Change data attribute constant to import from bezier-tokens
-    rootElement.setAttribute('data-bezier-theme', themeName)
-    return function cleanup() {
-      rootElement.removeAttribute('data-bezier-theme')
-    }
-  }, [
-    rootElement,
-    themeName,
-  ])
-
   return (
     <WindowProvider window={window}>
-      <TokenProvider themeName={themeName}>
-        <FeatureProvider features={features}>
-          <TooltipProvider>
-            { children }
-          </TooltipProvider>
-        </FeatureProvider>
-      </TokenProvider>
+      <FeatureProvider features={features}>
+        <TooltipProvider>
+          <ThemeProvider themeName={themeName}>
+            <div data-bezier-root>
+              { children }
+            </div>
+          </ThemeProvider>
+        </TooltipProvider>
+      </FeatureProvider>
     </WindowProvider>
   )
 }
